Fiddling with cache management

+++ lib/util/Job.cpp
@@ -40,6 +40,7 @@ bool sjef::util::Job::push_rundir(int verbosity) {
     command += " -v";
   m_project.m_trace(2 - verbosity) << "Push rsync: " << command << std::endl;
   auto start_time = std::chrono::steady_clock::now();
+  (*m_backend_command_server)("mkdir -p " + m_remote_cache_directory);
   Shell()(command, verbosity);
   if (verbosity > 1)
     m_project.m_trace(3 - verbosity)
@@ -50,6 +51,7 @@ bool sjef::util::Job::push_rundir(int verbosity) {
 }
 
 bool sjef::util::Job::pull_rundir(int verbosity) {
+  m_trace(3-verbosity) << "pull_rundir " << verbosity << std::endl;
   if (localhost())
     return true;
   std::string command = "rsync --archive --copy-links --timeout=5";
@@ -62,12 +64,13 @@ bool sjef::util::Job::pull_rundir(int verbosity) {
     command += " -v";
   m_project.m_trace(2 - verbosity) << "Pull rsync: " << command << std::endl;
   auto start_time = std::chrono::steady_clock::now();
-  Shell()(command, verbosity);
+  auto rsync_out = Shell()(command, verbosity);
   if (verbosity > 1)
     m_project.m_trace(3 - verbosity)
         << "time for pull_rundir() rsync "
         << std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::steady_clock::now() - start_time).count()
-        << "ms" << std::endl;
+        << "ms" << std::endl
+        << "Output from rsync\n:" << rsync_out << std::endl;
   return true; // TODO: implement more robust error checking
 }
 
@@ -93,22 +96,19 @@ std::string Job::run(const std::string& command, int verbosity, bool wait) {
   //  m_trace(4 - verbosity) << "Job::run substr=" << substr << " m_backend.run_command=" << m_backend.run_command
   //                         << std::endl;
   //  auto is_run_command = substr == m_backend.run_command;
-  auto is_run_command = true;
-  //  m_trace(4-verbosity)<< "is_run_command: "<<is_run_command<<std::endl;
-  if (is_run_command) {
-    m_initial_status = waiting;
-    m_job_number = 0; // pauses status polling
-                      //    std::cout << "Job::run() set initial status to waiting, and pause polling"<<std::endl;
-  }
+  m_initial_status = waiting;
+  m_job_number = 0; // pauses status polling
+                    //    std::cout << "Job::run() set initial status to waiting, and pause polling"<<std::endl;
   set_status(waiting);
-  auto backend_submits_batch = is_run_command and m_backend.run_jobnumber != "([0-9]+)";
-  if (backend_submits_batch)
+  auto backend_submits_batch = m_backend.run_jobnumber != "([0-9]+)";
+  if (!localhost())
     push_rundir(verbosity);
   m_trace(4 - verbosity) << "Job::run() gives directory " << m_project.filename("", "", 0) << std::endl;
   m_trace(4 - verbosity) << "before submit, m_backend_command_server? " << (m_backend_command_server == nullptr)
                          << std::endl;
-  auto run_output =
-      (*m_backend_command_server)(command, wait or backend_submits_batch, m_project.filename("", "", 0), verbosity);
+  auto run_output = (*m_backend_command_server)(
+      command, wait or backend_submits_batch,
+      localhost() ? m_project.filename("", "", 0).string() : m_remote_cache_directory, verbosity);
   if (backend_submits_batch) {
     std::smatch match;
     if (std::regex_search(run_output, match, std::regex{m_backend.run_jobnumber})) {
@@ -116,24 +116,15 @@ std::string Job::run(const std::string& command, int verbosity, bool wait) {
       m_job_number = std::stoi(match[1]);
       m_trace(4 - verbosity) << "Job::run backend_submits_batch m_job_number=" << m_job_number << std::endl;
     }
-  } else if (is_run_command) {
+  } else {
     m_trace(4 - verbosity) << "before job_number(), m_backend_command_server? " << (m_backend_command_server == nullptr)
                            << std::endl;
     m_job_number = m_backend_command_server->job_number();
     m_trace(4 - verbosity) << "Job::run is_run_command m_job_number=" << m_job_number << std::endl;
-  } else
-    m_trace(4 - verbosity) << "Job::run doesn't collect m_job_number" << std::endl;
-  //  if (is_run_command) {
-  //    m_closing=true;
-  //    m_poll_task.wait();
-  ////  }
-  ////  if (is_run_command) {
-  //    m_poll_task = std::async(std::launch::async, [this]() { this->poll_job(); });
-  //  }
-  //  sleep(1);
-  //  std::cout << "Job::run() returns " << run_output << std::endl;
+  }
   m_poll_task = std::async(std::launch::async, [this]() { this->poll_job(); });
   return run_output;
+  m_trace(2 - verbosity) << "Job::run() returns " << run_output << std::endl;
 }
 
 void Job::set_status(status stat) {
@@ -172,18 +163,19 @@ status Job::get_status(int verbosity) {
 }
 
 void Job::kill(int verbosity) {
-  m_trace(4-verbosity) << "Job::kill()"<<std::endl;
-//  std::cout << "Job::kill()"<<std::endl;
+  m_trace(4 - verbosity) << "Job::kill()" << std::endl;
+  //  std::cout << "Job::kill()"<<std::endl;
   {
     auto l = std::lock_guard(kill_mutex);
-//    std::cout << "Job::kill() gets mutex"<<std::endl;
-    auto status_string = (*m_backend_command_server)(m_backend.kill_command + " " + std::to_string(m_job_number),true,".",verbosity);
-//    std::cout << "Job::kill() finished killing"<<std::endl;
+    //    std::cout << "Job::kill() gets mutex"<<std::endl;
+    auto status_string =
+        (*m_backend_command_server)(m_backend.kill_command + " " + std::to_string(m_job_number), true, ".", verbosity);
+    //    std::cout << "Job::kill() finished killing"<<std::endl;
     set_status(killed);
-//    std::cout << "Job::kill() finished set_status()"<<std::endl;
+    //    std::cout << "Job::kill() finished set_status()"<<std::endl;
   }
   m_killed = true;
-//  std::cout << "Job::kill() set sentinel"<<std::endl;
+  //  std::cout << "Job::kill() set sentinel"<<std::endl;
 }
 void Job::poll_job(int verbosity) {
   using Clock = std::chrono::high_resolution_clock;
@@ -196,8 +188,8 @@ void Job::poll_job(int verbosity) {
     {
       auto l = std::lock_guard(kill_mutex);
       //      std::cout << "active polling cycle starts"<<std::endl;
-//      if (m_killed)
-//        std::cout << "poll_job received kill sentinel" << std::endl;
+      //      if (m_killed)
+      //        std::cout << "poll_job received kill sentinel" << std::endl;
 
       start = Clock::now();
       status = m_killed ? killed : get_status(verbosity);
@@ -207,7 +199,7 @@ void Job::poll_job(int verbosity) {
         if (m_initial_status == running or m_initial_status == completed or m_initial_status == waiting)
           status = completed;
       }
-      //    std::cout << "got status " << status << std::endl;
+      m_trace(4-verbosity) << "got status " << status << std::endl;
       pull_rundir(verbosity);
       set_status(status);
       //    std::cout << "set status " << m_project.status_message() << std::endl;
@@ -217,17 +209,27 @@ void Job::poll_job(int verbosity) {
         if (m_closing or status == completed or m_killed)
           break;
       }
-      //      std::cout << "active polling cycle stops"<<std::endl;
+            m_trace(4-verbosity) << "active polling cycle stops"<<std::endl;
     }
     using namespace std::literals::chrono_literals;
     std::this_thread::sleep_for(10ms);
     std::this_thread::sleep_for((stop - start) * 2);
   }
-  if (m_backend_command_server != nullptr and (status == completed or status == killed)) {
-    (*m_backend_command_server)("rm -rf " + m_remote_cache_directory);
+  if (!localhost() and m_backend_command_server != nullptr and (status == completed or status == killed)) {
+    m_trace(4-verbosity) << "Pull run directory at end of job " << std::endl;
+    m_trace(4-verbosity) << Shell()("echo local rundir;ls -lta " + m_project.filename("", "", 0).string()) << std::endl;
+    m_trace(4-verbosity) << "remote cache directory: " << m_remote_cache_directory << std::endl;
+    m_trace(4-verbosity) << (*m_backend_command_server)("echo remote cache;ls -lta " + m_remote_cache_directory + "  2>&1")
+              << std::endl;
+    pull_rundir(verbosity); // TODO don't remove unless it worked!
+    m_trace(4-verbosity) << Shell()("echo local rundir;ls -lta " + m_project.filename("", "", 0).string()) << std::endl;
+    m_trace(4-verbosity) << (*m_backend_command_server)("echo remote cache;ls -lta " + m_remote_cache_directory + "  2>&1")
+              << std::endl;
+    m_trace(4-verbosity) << "remove run directory " + m_remote_cache_directory + " at end of job " << std::endl;
+        (*m_backend_command_server)("rm -rf " + m_remote_cache_directory);
   }
   m_backend_command_server.reset(); // close down backend server as no longer needed
-  //  std::cout << "Polling stops" << std::endl;
+  //  m_trace(4-verbosity) << "Polling stops" << std::endl;
 }
 
 } // namespace sjef::util
